fix invalid use of subprojects
The keycodemapdb Meson subproject provides a program and a source input.
Since it is a subproject, Meson wants to sandbox that and requires it to
be explicitly exported. But this never happened -- instead, we manually
poked at files using the actual string path "subprojects/......"

This was always a Meson sandbox violation, but Meson 0.63.0 started
noticing it and erroring out.

Instead, do the right thing. Update the subproject to a version that has
a meson.build with actually meaningful contents -- namely, a files
variable and a found program. Then use these in order to run the needed
custom_target.

In the process, it is also necessary to correct the argument ordering
when running keymap-gen.

Reviewed-by: Marc-André Lureau <marcandre.lureau@redhat.com>